program-indicators-legends: add LegendSetHandlerShould
