+++ core/src/test/resources/legendset/legend.json
