+++ core/src/test/resources/legendset/legendSet.json
